chore: renaming DocumentSignerFlow and exposing it's inner blocks
