+++ src/components/Company/DocumentSigner/documentSignerStateMachine.tsx
@@ -3,7 +3,7 @@ import type { Signatory } from '@gusto/embedded-api/models/components/signatory'
 import { AssignSignatory as BlockAssignSignatory } from '../AssignSignatory'
 import { DocumentList as BlockDocumentList } from './DocumentList'
 import { SignatureForm as BlockSignatureForm } from './SignatureForm'
-import type { DocumentSignerContextInterface } from './useDocumentSignerFlow'
+import type { DocumentSignerContextInterface } from './useDocumentSigner'
 import type { companyEvents } from '@/shared/constants'
 import { useFlow } from '@/components/Flow/useFlow'
 import { ensureRequired } from '@/helpers/ensureRequired'
