+++ src/components/Company/DocumentSigner/index.ts
@@ -1,0 +1,1 @@
+export { DocumentSigner } from './DocumentSigner'
