+++ src/components/Company/DocumentSigner/stateMachine.ts
@@ -2,7 +2,7 @@ import { state, transition, reduce } from 'robot3'
 import type { EventPayloads } from './documentSignerStateMachine'
 import { SignatureForm, AssignSignatory, DocumentList } from './documentSignerStateMachine'
 import { assignSignatoryState } from './assignSignatoryState'
-import type { DocumentSignerContextInterface } from './useDocumentSignerFlow'
+import type { DocumentSignerContextInterface } from './useDocumentSigner'
 import type { MachineEventType } from '@/types/Helpers'
 import { companyEvents } from '@/shared/constants'
