--- src/components/Company/DocumentSignerFlow/index.ts
@@ -1,1 +1,0 @@
-export { DocumentSignerFlow } from './DocumentSignerFlow'
