+++ src/components/Company/index.tsx
@@ -1,8 +1,8 @@
 export { Industry } from './Industry'
 export { AssignSignatory } from './AssignSignatory'
-export { DocumentList } from './DocumentSignerFlow/DocumentList'
-export { SignatureForm } from './DocumentSignerFlow/SignatureForm'
-export { DocumentSignerFlow } from './DocumentSignerFlow'
+export { DocumentList } from './DocumentSigner/DocumentList'
+export { SignatureForm } from './DocumentSigner/SignatureForm'
+export { DocumentSigner } from './DocumentSigner'
 export { OnboardingOverview } from './OnboardingOverview/OnboardingOverview'
 export { LocationsFlow } from './Locations/LocationsFlow'
 export { PaySchedule } from './PaySchedule'
