+++ src/components/Employee/DocumentSigner/DocumentSigner.tsx
@@ -4,14 +4,16 @@ import {
   type DocumentSignerContextInterface,
 } from './documentSignerStateMachine'
 import { documentSignerMachine } from './stateMachine'
+import { DocumentList } from './DocumentList/DocumentList'
+import { SignatureForm } from './SignatureForm/SignatureForm'
 import { Flow } from '@/components/Flow/Flow'
 import type { BaseComponentInterface } from '@/components/Base'
 
-export interface DocumentSignerFlowProps extends BaseComponentInterface {
+export interface DocumentSignerProps extends BaseComponentInterface {
   employeeId: string
 }
 
-export const DocumentSignerFlow = ({ employeeId, onEvent }: DocumentSignerFlowProps) => {
+export const DocumentSigner = ({ employeeId, onEvent }: DocumentSignerProps) => {
   const documentSigner = createMachine(
     'index',
     documentSignerMachine,
@@ -23,3 +25,6 @@ export const DocumentSignerFlow = ({ employeeId, onEvent }: DocumentSignerFlowPr
   )
   return <Flow machine={documentSigner} onEvent={onEvent} />
 }
+
+DocumentSigner.DocumentList = DocumentList
+DocumentSigner.SignatureForm = SignatureForm
