+++ src/components/Employee/DocumentSigner/index.ts
@@ -1,1 +1,1 @@
-export * from './DocumentSignerFlow'
+export * from './DocumentSigner'
