+++ src/components/Employee/index.ts
@@ -6,6 +6,6 @@ export { Compensation } from './Compensation'
 export { Taxes } from './Taxes'
 export { PaymentMethod } from './PaymentMethod'
 export { Landing } from './Landing'
-export { DocumentSignerFlow } from './DocumentSigner'
+export { DocumentSigner } from './DocumentSigner'
 export { OnboardingFlow } from './OnboardingFlow/OnboardingFlow'
 export { SelfOnboardingFlow } from './SelfOnboardingFlow/SelfOnboardingFlow'
